pyproject.toml
[project]
name = "chess-transformer"
version = "0.1.0"
description = "Chess model via RandOpt post-training with AttnRes on a pretrained LLM"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "transformers>=4.40",
    "accelerate>=0.28",
    "datasets>=2.18",
    "python-chess>=1.10",
    "numpy>=1.26",
    "tqdm",
    "safetensors",
    "huggingface-hub",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "ipython",
]
stockfish = [
    "stockfish",  # python wrapper for stockfish engine
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py310"
